clean account csv to json
