+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.1.5+26",
+  "version": "0.1.5+27",
   "private": false,
   "scripts": {
     "dev": "next dev",
