+++ src/constants/account.ts
@@ -63,3 +63,5 @@ export const EVENT_TYPE_TO_VOUCHER_TYPE_MAP: {
   [EventType.PAYMENT]: VoucherType.EXPENSE,
   [EventType.TRANSFER]: VoucherType.TRANSFER,
 };
+
+export const MISSING_CODE_MARKERS = ['!', '@', '#', '$', '%'];
